+++ controllers/DocFlow_Controller.php
@@ -339,11 +339,20 @@ class DocFlow_Controller extends ControllerSQL{
 						$ar['file_id'].'.sig';		
 				}
 				
-				if($cur_sig_exists)unlink($cur_sig);
-				rename($prev_sig,$new_cur_sig);
+				if (file_exists($prev_sig)){
+					if($cur_sig_exists)unlink($cur_sig);
+					rename($prev_sig,$new_cur_sig);
+				}
 			
 				$pki_man = new PKIManager(PKI_PATH,PKI_CRL_VALIDITY,PKI_MODE);
-				pki_log_sig_check($new_cur_sig, $file_doc, $this->getExtDbVal($pm,'file_id'), $pki_man, $this->getDbLinkMaster());
+				$verif_res = pki_log_sig_check($new_cur_sig, $file_doc, $this->getExtDbVal($pm,'file_id'), $pki_man, $this->getDbLinkMaster());
+				if (
+					!$verif_res->checkPassed
+					&&
+					( PKI_SIG_ERROR=='ALL' || (PKI_SIG_ERROR=='NO_CERT' && !count($verif_res->signatures) ) )
+				){
+					throw new Exception(sprintf(ER_VERIF_SIG,$verif_res->checkError));
+				}		
 			}
 			
 			$this->getDbLinkMaster()->query("COMMIT");
